Removed functionality from the action handler to help with testing.
