+++ HousingSearchListener.Tests/Gateways/PersonMessageFactoryTests.cs
@@ -1,4 +1,5 @@
-﻿using Amazon.Lambda.SNSEvents;
+﻿using System;
+using Amazon.Lambda.SNSEvents;
 using FluentAssertions;
 using HousingSearchListener.Gateways;
 using Xunit;
@@ -18,6 +19,8 @@ namespace HousingSearchListener.Tests.Gateways
         public void GivenAnSnsRecord_WhenProcessed_ShouldDeserializeCorrectly()
         {
             // given
+            Environment.SetEnvironmentVariable("PersonApiUrl", "http://127.0.0.1");
+            Environment.SetEnvironmentVariable("PersonApiToken", "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjEwNSIsImNvbnN1bWVyTmFtZSI6IlRlbmFudHMgYW5kIExlYXNlaG9sZGVycyIsImNvbnN1bWVyVHlwZSI6IjIiLCJuYmYiOjE2MjEzMzAxODcsImV4cCI6MTkzNjg2Mjk4NywiaWF0IjoxNjIxMzMwMTg3fQ.cK37Miiu3huq_lHHn0kyfdOvAqXCO0cyXquiNOiK318");
             var message =
                 "{ \"id\": \"8e648f3d-9556-4896-8400-211cb1c5451b\", \"eventType\": \"PersonCreatedEvent\", \"sourceDomain\": \"Person\", \"sourceSystem\": " +
                 "\"PersonAPI\", \"version\": \"v1\", \"correlationId\": \"f4d541d0-7c07-4524-8296-2d0d50cb58f4\", \"dateTime\": \"2021-05-17T11:59:57.25Z\", " +
