+++ HousingSearchListener/V1/Interfaces/IElasticSearchHelper.cs
@@ -8,19 +8,4 @@ namespace HousingSearchListener.V1.Interfaces
     {
         Task<CreateResponse> Create(ESPerson esPerson);
     }
-
-    public class ElasticSearchHelper : IElasticSearchHelper
-    {
-        private readonly IElasticClient _elasticClient;
-
-        public ElasticSearchHelper(IElasticClient elasticClient)
-        {
-            _elasticClient = elasticClient;
-        }
-
-        public async Task<CreateResponse> Create(ESPerson esPerson)
-        {
-            return await _elasticClient.CreateAsync<ESPerson>(new CreateRequest<ESPerson>(esPerson));
-        }
-    }
 }
