+++ HousingSearchListener/V1/Interfaces/IElasticSearchUpdater.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using Amazon.Lambda.SNSEvents;
+
+namespace HousingSearchListener
+{
+    public interface IElasticSearchUpdater
+    {
+        Task Update(SNSEvent snsEvent);
+    }
+}
